Moving worker pool to own dir
