+++ src/ray/raylet/local_object_manager.h
@@ -24,7 +24,7 @@
 #include "ray/object_manager/common.h"
 #include "ray/object_manager/object_directory.h"
 #include "ray/pubsub/subscriber.h"
-#include "ray/raylet/worker_pool.h"
+#include "ray/raylet/worker_pool/worker_pool.h"
 #include "ray/rpc/worker/core_worker_client_pool.h"
 #include "ray/util/util.h"
 #include "src/ray/protobuf/node_manager.pb.h"
