+++ src/ray/raylet/worker_killing_policy.h
@@ -20,7 +20,7 @@
 #include "ray/common/asio/periodical_runner.h"
 #include "ray/common/memory_monitor.h"
 #include "ray/raylet/worker.h"
-#include "ray/raylet/worker_pool.h"
+#include "ray/raylet/worker_pool/worker_pool.h"
 
 namespace ray {
